refactor(protocol-designer): initial refactoring of MoveLiquidTools
create new two components to reduce MoveLiquidTools and add basic tests

close AUTH-1417
