+++ protocol-designer/src/organisms/TipPositionModal/index.tsx
@@ -27,6 +27,7 @@ import { TipPositionSideView } from './TipPositionSideView'
 
 import type { ChangeEvent } from 'react'
 import type { StepFieldName } from '../../form-types'
+import type { PrefixType } from '../../resources/types'
 
 type Offset = 'x' | 'y' | 'z'
 interface PositionSpec {
@@ -43,7 +44,7 @@ interface TipPositionModalProps {
   wellXWidthMm: number
   wellYWidthMm: number
   isIndeterminate?: boolean
-  prefix: 'aspirate' | 'dispense' | 'mix'
+  prefix: PrefixType
 }
 
 export function TipPositionModal(
