+++ protocol-designer/src/organisms/WellOrderModal/index.tsx
@@ -17,7 +17,9 @@ import {
 import { LINK_BUTTON_STYLE } from '../../atoms'
 import { getMainPagePortalEl } from '../Portal'
 import { WellOrderVisualization } from './WellOrderVisualization'
+
 import type { WellOrderOption } from '../../form-types'
+import type { PrefixType } from '../../resources/types'
 
 const DEFAULT_FIRST: WellOrderOption = 't2b'
 const DEFAULT_SECOND: WellOrderOption = 'l2r'
@@ -31,7 +33,7 @@ const WELL_ORDER_VALUES: WellOrderOption[] = [
 export interface WellOrderModalProps {
   isOpen: boolean
   closeModal: () => void
-  prefix: 'aspirate' | 'dispense' | 'mix'
+  prefix: PrefixType
   firstName: string
   secondName: string
   firstValue?: WellOrderOption | null
