+++ protocol-designer/src/pages/Designer/ProtocolSteps/StepForm/PipetteFields/FlowRateField.tsx
@@ -5,10 +5,12 @@ import { SPACING } from '@opentrons/components'
 import { selectors as stepFormSelectors } from '../../../../../step-forms'
 import { getMatchingTipLiquidSpecs } from '../../../../../utils'
 import { InputStepFormField } from '../../../../../molecules'
+
 import type { FieldProps } from '../types'
+import type { FlowRateType } from '../../../../../resources/types'
 
 interface FlowRateFieldProps extends FieldProps {
-  flowRateType: 'aspirate' | 'dispense' | 'blowout'
+  flowRateType: FlowRateType
   volume: unknown
   tiprack: unknown
   pipetteId?: string | null
