+++ protocol-designer/src/pages/Designer/ProtocolSteps/StepForm/PipetteFields/WellsOrderField.tsx
@@ -13,11 +13,13 @@ import {
   ALIGN_CENTER,
 } from '@opentrons/components'
 import { WellOrderModal } from '../../../../../organisms'
-import type { WellOrderOption } from '../../../../../form-types'
+
 import type { FieldProps } from '../types'
+import type { WellOrderOption } from '../../../../../form-types'
+import type { PrefixType } from '../../../../../resources/types'
 
 export interface WellsOrderFieldProps {
-  prefix: 'aspirate' | 'dispense' | 'mix'
+  prefix: PrefixType
   firstName: string
   secondName: string
   updateFirstWellOrder: FieldProps['updateValue']
