+++ protocol-designer/src/resources/types.ts
@@ -1,0 +1,3 @@
+export type PrefixType = 'aspirate' | 'dispense' | 'mix'
+
+export type FlowRateType = 'aspirate' | 'dispense' | 'blowout'
